+++ pandasdmx/__init__.py
@@ -29,7 +29,7 @@ def _init_logger():
         '%(asctime)s %(name)s - %(levelname)s: %(message)s')
     handler.setFormatter(fmt)
     logger.addHandler(handler)
-    logger.setLevel(logging.INFO)
+    logger.setLevel(logging.ERROR)
     return logger
 
 logger = _init_logger()