+++ pandasdmx/remote.py
@@ -52,7 +52,7 @@ class REST:
         if cache:
             requests_cache.install_cache(**cache)
 
-    def get(self, url, fromfile=None, params={}):
+    def get(self, url, fromfile=None, params={}, headers={}):
         '''Get SDMX message from REST service or local file
 
         Args:
@@ -63,6 +63,8 @@ class REST:
             fromfile(str): path to SDMX file containing an SDMX message.
                 It will be passed on to the
                 reader for parsing.
+            headers(dict): http headers. Overwrite instance-wide headers.
+                Default is {}
 
         Returns:
             tuple: three objects:
@@ -84,13 +86,13 @@ class REST:
             except TypeError:
                 # so fromfile must be file-like
                 source = fromfile
-            final_url = headers = status_code = None
+            final_url = resp_headers = status_code = None
         else:
-            source, final_url, headers, status_code = self.request(
-                url, params=params)
-        return source, final_url, headers, status_code
+            source, final_url, resp_headers, status_code = self.request(
+                url, params=params, headers=headers)
+        return source, final_url, resp_headers, status_code
 
-    def request(self, url, params={}):
+    def request(self, url, params={}, headers={}):
         """
         Retrieve SDMX messages.
         If needed, override in subclasses to support other data providers.
@@ -99,8 +101,15 @@ class REST:
         :type url: str
         :return: the xml data as file-like object
         """
+        # Generate current config. Merge in any given headers
+        cur_config = self.config.copy()
+        if 'headers' in cur_config:
+            cur_config['headers'] = cur_config['headers'].copy()
+            cur_config['headers'].update(headers)
+        else:
+            cur_config['headers'] = headers
 
-        with closing(requests.get(url, params=params, **self.config)) as response:
+        with closing(requests.get(url, params=params, **cur_config)) as response:
             if response.status_code == requests.codes.OK:
                 source = STF(max_size=self.max_size)
                 for c in response.iter_content(chunk_size=1000000):